database/requirements.txt
psycopg2-binary>=2.9.3
python-dotenv>=0.20.0
faker>=13.3.4